Allows for un-caputred messages to checked and added to playlists
Fixes issue where if spotbot gets knocked offline the user has to re-add all messages that have been sent into the channel. No longer, now by setting the flag "check_past_on_boot" in setup.json to true. The bot will now look for messages that do not have a spotbot emoji associated with it.

Need to integrate with @karltrowbridge1 database to verify if the song has been sent or not